Refactor test location
